move genPodJs from bin to rollup-plugin-gen-podjs
